fix direct io asan bug
